+++ Tests/Unit/Domain/Model/GermanZipCodeTest.php
@@ -9,6 +9,9 @@ use OliverKlee\Oelib\Domain\Model\GermanZipCode;
 use OliverKlee\Oelib\Interfaces\Geo;
 use TYPO3\CMS\Extbase\DomainObject\AbstractEntity;
 
+/**
+ * @covers \OliverKlee\Oelib\Domain\Model\GermanZipCode
+ */
 class GermanZipCodeTest extends UnitTestCase
 {
     /**