+++ Tests/Unit/Geocoding/Fixtures/TestingGeo.php
@@ -13,8 +13,6 @@ use OliverKlee\Oelib\Model\AbstractModel;
 class TestingGeo extends AbstractModel implements Geo
 {
     /**
-     * whether this object has a geo error
-     *
      * @var bool
      */
     private $hasGeoError = false;
@@ -25,25 +23,21 @@ class TestingGeo extends AbstractModel implements Geo
     private $geoErrorReason = '';
 
     /**
-     * the address of this object
-     *
      * @var string
      */
     private $address = '';
 
     /**
-     * the geo coordinates of this object
-     *
-     * @var array{latitude?: float, longitude?: float}
+     * @var array{latitude: float, longitude: float}|null
      */
-    private $coordinates = [];
+    private $coordinates = null;
 
     /**
-     * Returns this object's address formatted for a geo lookup, for example
+     * Returns this object's address formatted for a geocoding lookup, for example
      * "Pariser Str. 50, 53117 Auerberg, Bonn, DE". Any part of this address
      * might be missing, though.
      *
-     * @return string this object's address formatted for a geo lookup,
+     * @return string this object's address formatted for a geocoding lookup,
      *                will be empty if this object has no address
      */
     public function getGeoAddress(): string
@@ -52,8 +46,6 @@ class TestingGeo extends AbstractModel implements Geo
     }
 
     /**
-     * Sets this object's geo address.
-     *
      * @param string $address the address to set, for example
      *        "Pariser Str. 50, 53117 Auerberg, Bonn, DE", may be empty
      */
@@ -62,13 +54,6 @@ class TestingGeo extends AbstractModel implements Geo
         $this->address = $address;
     }
 
-    /**
-     * Checks whether this object has a non-empty address suitable for a geo
-     * lookup.
-     *
-     * @return bool TRUE if this object has a non-empty address, FALSE
-     *                 otherwise
-     */
     public function hasGeoAddress(): bool
     {
         return $this->address !== '';
@@ -81,7 +66,7 @@ class TestingGeo extends AbstractModel implements Geo
      */
     public function getGeoCoordinates(): array
     {
-        if (!isset($this->coordinates['latitude'], $this->coordinates['longitude'])) {
+        if (!\is_array($this->coordinates)) {
             throw new \BadMethodCallException('Missing geo coordinates!', 1633018227);
         }
 
@@ -89,8 +74,6 @@ class TestingGeo extends AbstractModel implements Geo
     }
 
     /**
-     * Sets this object's coordinates.
-     *
      * @param array{latitude: float, longitude: float} $coordinates
      */
     public function setGeoCoordinates(array $coordinates): void
@@ -100,34 +83,27 @@ class TestingGeo extends AbstractModel implements Geo
 
     /**
      * Checks whether this object has non-empty coordinates.
-     *
-     * @return bool TRUE if this object has both a non-empty longitude and
-     *                 a non-empty latitude, FALSE otherwise
      */
     public function hasGeoCoordinates(): bool
     {
-        return $this->coordinates !== [];
+        return \is_array($this->coordinates);
     }
 
     /**
      * Purges this object's geo coordinates.
      *
-     * Note: Calling this function has no influence on this object's geo error
-     * status.
+     * Note: Calling this function has no influence on this object's geo error status.
      */
     public function clearGeoCoordinates(): void
     {
-        $this->coordinates = [];
+        $this->coordinates = null;
     }
 
     /**
-     * Checks whether there has been a problem with this object's geo
-     * coordinates.
+     * Checks whether there has been a problem with this object's geo coordinates.
      *
      * Note: This function only checks whether there has been an error with the
      * coordinates, not whether this object actually has coordinates.
-     *
-     * @return bool TRUE if there has been an error, FALSE otherwise
      */
     public function hasGeoError(): bool
     {
@@ -143,9 +119,6 @@ class TestingGeo extends AbstractModel implements Geo
         $this->geoErrorReason = $reason;
     }
 
-    /**
-     * @return string
-     */
     public function getGeoErrorReason(): string
     {
         return $this->geoErrorReason;